IDCODE: IDCODE[11:1] are ManufID, maybe this will be a good choice ?
ManufID[6:0] are derived from the last byte of the
JEDEC code with the parity bit discarded.
ManufID[10:7] provide a binary count of the number
of bytes in the JEDEC code that contain the
continuation character (0x7F). When the number of
continuations characters exceeds 15, these four bits
contain the modulo-16 count of the number of
continuation characters.